@@ -4,25 +4,47 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.6.2",
+    "@ant-design/plots": "^2.4.0",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@mantine/core": "^8.0.2",
+    "@mantine/hooks": "^8.0.2",
+    "@radix-ui/react-avatar": "^1.1.10",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@sendbird/uikit-react": "^3.16.11",
+    "@tabler/icons-react": "^3.33.0",
+    "antd": "^4.16.6",
+    "apexcharts": "^4.7.0",
+    "axios": "^1.10.0",
+    "cross-env": "^7.0.3",
+    "dayjs": "^1.11.13",
+    "firebase": "^11.10.0",
+    "lucide-react": "^0.525.0",
+    "react": "^18.3.1",
+    "react-apexcharts": "^1.3.9",
+    "react-calendar": "^6.0.0",
+    "react-chartjs-2": "^5.3.0",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.5.0",
+    "react-quill": "^2.0.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^5.0.1",
+    "styled-components": "^5.3.0",
+    "web-vitals": "^1.1.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.16.0",
+    "@craco/craco": "^7.1.0",
+    "@openapitools/openapi-generator-cli": "^2.21.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "antd": "^4.16.6",
-    "apexcharts": "^3.27.1",
-    "install": "^0.13.0",
-    "react": "^17.0.2",
-    "react-apexcharts": "^1.3.9",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.0",
-    "web-vitals": "^1.1.2",
-    "cross-env": "^7.0.3"
+    "@testing-library/user-event": "^12.8.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "cross-env CI=false cross-env NODE_OPTIONS=--openssl-legacy-provider react-scripts build",
-    "test": "react-scripts test",
+    "start": "cross-env NODE_OPTIONS=--openssl-legacy-provider craco start",
+    "build": "cross-env NODE_OPTIONS=--openssl-legacy-provider craco build",
+    "test": "craco test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -42,8 +64,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@babel/core": "^7.16.0"
   }
 }
