@@ -1,14 +1,4 @@
-/*!
-  =========================================================
-  * Muse Ant Design Dashboard - v1.0.0
-  =========================================================
-  * Product Page: https://www.creative-tim.com/product/muse-ant-design-dashboard
-  * Copyright 2021 Creative Tim (https://www.creative-tim.com)
-  * Licensed under MIT (https://github.com/creativetimofficial/muse-ant-design-dashboard/blob/main/LICENSE.md)
-  * Coded by Creative Tim
-  =========================================================
-  * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-*/
+import React from "react";
 import { useState } from "react";
 
 import {
@@ -45,7 +35,7 @@ import project3 from "../assets/images/home-decor-3.jpeg";
 
 function Profile() {
   const [imageURL, setImageURL] = useState(false);
-  const [, setLoading] = useState(false);
+  const [loading, setLoading] = useState(false);
 
   const getBase64 = (img, callback) => {
     const reader = new FileReader();
@@ -56,24 +46,24 @@ function Profile() {
   const beforeUpload = (file) => {
     const isJpgOrPng = file.type === "image/jpeg" || file.type === "image/png";
     if (!isJpgOrPng) {
-      message.error("You can only upload JPG/PNG file!");
+      message.error("Bạn chỉ có thể tải lên file JPG/PNG!");
     }
     const isLt2M = file.size / 1024 / 1024 < 2;
     if (!isLt2M) {
-      message.error("Image must smaller than 2MB!");
+      message.error("Kích thước ảnh phải nhỏ hơn 2MB!");
     }
     return isJpgOrPng && isLt2M;
   };
 
   const handleChange = (info) => {
     if (info.file.status === "uploading") {
-      setLoading(false);
+      setLoading(true);
       return;
     }
     if (info.file.status === "done") {
       getBase64(info.file.originFileObj, (imageUrl) => {
         setLoading(false);
-        setImageURL(false);
+        setImageURL(imageUrl);
       });
     }
   };
@@ -101,7 +91,7 @@ function Profile() {
   const uploadButton = (
     <div className="ant-upload-text font-semibold text-dark">
       {<VerticalAlignTopOutlined style={{ width: 20, color: "#000" }} />}
-      <div>Upload New Project</div>
+      <div>Tải Ảnh Lên</div>
     </div>
   );
 
@@ -109,267 +99,247 @@ function Profile() {
     {
       title: "Sophie B.",
       avatar: convesionImg,
-      description: "Hi! I need more information…",
+      description: "Dịch vụ rất tốt, nhân viên chuyên nghiệp",
     },
     {
       title: "Anne Marie",
       avatar: convesionImg2,
-      description: "Awesome work, can you…",
+      description: "Tôi rất hài lòng với kết quả",
     },
     {
       title: "Ivan",
       avatar: convesionImg3,
-      description: "About files I can…",
+      description: "Sẽ tiếp tục sử dụng dịch vụ",
     },
     {
       title: "Peterson",
       avatar: convesionImg4,
-      description: "Have a great afternoon…",
+      description: "Chất lượng dịch vụ tuyệt vời",
     },
     {
-      title: "Nick Daniel",
+      title: "Maria",
       avatar: convesionImg5,
-      description: "Hi! I need more information…",
+      description: "Đội ngũ nhân viên rất nhiệt tình",
     },
   ];
 
   const project = [
     {
       img: project1,
-      titlesub: "Project #1",
-      title: "Modern",
+      titlesub: "Thiết Kế #1",
+      title: "Thiết Kế Sân Vườn Hiện Đại",
       disciption:
-        "As Uber works through a huge amount of internal management turmoil.",
+        "Dự án thiết kế và thi công sân vườn theo phong cách hiện đại cho biệt thự",
     },
     {
       img: project2,
-      titlesub: "Project #2",
-      title: "Scandinavian",
+      titlesub: "Thiết Kế #2",
+      title: "Vườn Trên Sân Thượng",
       disciption:
-        "Music is something that every person has his or her own specific opinion about.",
+        "Thiết kế không gian xanh trên sân thượng kết hợp với khu vực thư giãn",
     },
     {
       img: project3,
-      titlesub: "Project #3",
-      title: "Minimalist",
+      titlesub: "Thiết Kế #3",
+      title: "Sân Vườn Phong Cách Nhật",
       disciption:
-        "Different people have different taste, and various types of music, Zimbali Resort",
+        "Thiết kế sân vườn theo phong cách Nhật Bản với hồ cá Koi và cây cảnh",
     },
   ];
 
-  return (
-    <>
-      <div
-        className="profile-nav-bg"
-        style={{ backgroundImage: "url(" + BgProfile + ")" }}
-      ></div>
+  // return (
+  //   <>
+  //     <div
+  //       className="profile-nav-bg"
+  //       style={{ backgroundImage: "url(" + BgProfile + ")" }}
+  //     ></div>
 
-      <Card
-        className="card-profile-head"
-        bodyStyle={{ display: "none" }}
-        title={
-          <Row justify="space-between" align="middle" gutter={[24, 0]}>
-            <Col span={24} md={12} className="col-info">
-              <Avatar.Group>
-                <Avatar size={74} shape="square" src={profilavatar} />
+  //     <Card
+  //       className="card-profile-head"
+  //       bodyStyle={{ display: "none" }}
+  //       title={
+  //         <Row justify="space-between" align="middle" gutter={[24, 0]}>
+  //           <Col span={24} md={12} className="col-info">
+  //             <Avatar.Group>
+  //               <Avatar size={74} shape="square" src={profilavatar} />
 
-                <div className="avatar-info">
-                  <h4 className="font-semibold m-0">Sarah Jacob</h4>
-                  <p>CEO / Co-Founder</p>
-                </div>
-              </Avatar.Group>
-            </Col>
-            <Col
-              span={24}
-              md={12}
-              style={{
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "flex-end",
-              }}
-            >
-              <Radio.Group defaultValue="a">
-                <Radio.Button value="a">OVERVIEW</Radio.Button>
-                <Radio.Button value="b">TEAMS</Radio.Button>
-                <Radio.Button value="c">PROJECTS</Radio.Button>
-              </Radio.Group>
-            </Col>
-          </Row>
-        }
-      ></Card>
+  //               <div className="avatar-info">
+  //                 <h4 className="font-semibold m-0">Nguyễn Văn A</h4>
+  //                 <p>Quản Lý</p>
+  //               </div>
+  //             </Avatar.Group>
+  //           </Col>
+  //           <Col
+  //             span={24}
+  //             md={12}
+  //             style={{
+  //               display: "flex",
+  //               alignItems: "center",
+  //               justifyContent: "flex-end",
+  //             }}
+  //           >
+  //             <Radio.Group defaultValue="a">
+  //               <Radio.Button value="a">THÔNG TIN</Radio.Button>
+  //               <Radio.Button value="b">NHÓM</Radio.Button>
+  //               <Radio.Button value="c">DỰ ÁN</Radio.Button>
+  //             </Radio.Group>
+  //           </Col>
+  //         </Row>
+  //       }
+  //     ></Card>
 
-      <Row gutter={[24, 0]}>
-        <Col span={24} md={8} className="mb-24 ">
-          <Card
-            bordered={false}
-            className="header-solid h-full"
-            title={<h6 className="font-semibold m-0">Platform Settings</h6>}
-          >
-            <ul className="list settings-list">
-              <li>
-                <h6 className="list-header text-sm text-muted">ACCOUNT</h6>
-              </li>
-              <li>
-                <Switch defaultChecked />
-
-                <span>Email me when someone follows me</span>
-              </li>
-              <li>
-                <Switch />
-                <span>Email me when someone answers me</span>
-              </li>
-              <li>
-                <Switch defaultChecked />
-                <span>Email me when someone mentions me</span>
-              </li>
-              <li>
-                <h6 className="list-header text-sm text-muted m-0">
-                  APPLICATION
-                </h6>
-              </li>
-              <li>
-                <Switch defaultChecked />
-                <span>New launches and projects</span>
-              </li>
-              <li>
-                <Switch defaultChecked />
-                <span>Monthly product updates</span>
-              </li>
-              <li>
-                <Switch defaultChecked />
-                <span>Subscribe to newsletter</span>
-              </li>
-            </ul>
-          </Card>
-        </Col>
-        <Col span={24} md={8} className="mb-24">
-          <Card
-            bordered={false}
-            title={<h6 className="font-semibold m-0">Profile Information</h6>}
-            className="header-solid h-full card-profile-information"
-            extra={<Button type="link">{pencil}</Button>}
-            bodyStyle={{ paddingTop: 0, paddingBottom: 16 }}
-          >
-            <p className="text-dark">
-              {" "}
-              Hi, I’m Alec Thompson, Decisions: If you can’t decide, the answer
-              is no. If two equally difficult paths, choose the one more painful
-              in the short term (pain avoidance is creating an illusion of
-              equality).{" "}
-            </p>
-            <hr className="my-25" />
-            <Descriptions title="Oliver Liam">
-              <Descriptions.Item label="Full Name" span={3}>
-                Sarah Emily Jacob
-              </Descriptions.Item>
-              <Descriptions.Item label="Mobile" span={3}>
-                (44) 123 1234 123
-              </Descriptions.Item>
-              <Descriptions.Item label="Email" span={3}>
-                sarahjacob@mail.com
-              </Descriptions.Item>
-              <Descriptions.Item label="Location" span={3}>
-                USA
-              </Descriptions.Item>
-              <Descriptions.Item label="Social" span={3}>
-                <a href="#pablo" className="mx-5 px-5">
-                  {<TwitterOutlined />}
-                </a>
-                <a href="#pablo" className="mx-5 px-5">
-                  {<FacebookOutlined style={{ color: "#344e86" }} />}
-                </a>
-                <a href="#pablo" className="mx-5 px-5">
-                  {<InstagramOutlined style={{ color: "#e1306c" }} />}
-                </a>
-              </Descriptions.Item>
-            </Descriptions>
-          </Card>
-        </Col>
-        <Col span={24} md={8} className="mb-24">
-          <Card
-            bordered={false}
-            title={<h6 className="font-semibold m-0">Conversations</h6>}
-            className="header-solid h-full"
-            bodyStyle={{ paddingTop: 0, paddingBottom: 16 }}
-          >
-            <List
-              itemLayout="horizontal"
-              dataSource={data}
-              split={false}
-              className="conversations-list"
-              renderItem={(item) => (
-                <List.Item actions={[<Button type="link">REPLY</Button>]}>
-                  <List.Item.Meta
-                    avatar={
-                      <Avatar shape="square" size={48} src={item.avatar} />
-                    }
-                    title={item.title}
-                    description={item.description}
-                  />
-                </List.Item>
-              )}
-            />
-          </Card>
-        </Col>
-      </Row>
-      <Card
-        bordered={false}
-        className="header-solid mb-24"
-        title={
-          <>
-            <h6 className="font-semibold">Projects</h6>
-            <p>Architects design houses</p>
-          </>
-        }
-      >
-        <Row gutter={[24, 24]}>
-          {project.map((p, index) => (
-            <Col span={24} md={12} xl={6} key={index}>
-              <Card
-                bordered={false}
-                className="card-project"
-                cover={<img alt="example" src={p.img} />}
-              >
-                <div className="card-tag">{p.titlesub}</div>
-                <h5>{p.titile}</h5>
-                <p>{p.disciption}</p>
-                <Row gutter={[6, 0]} className="card-footer">
-                  <Col span={12}>
-                    <Button type="button">VIEW PROJECT</Button>
-                  </Col>
-                  <Col span={12} className="text-right">
-                    <Avatar.Group className="avatar-chips">
-                      <Avatar size="small" src={profilavatar} />
-                      <Avatar size="small" src={convesionImg} />
-                      <Avatar size="small" src={convesionImg2} />
-                      <Avatar size="small" src={convesionImg3} />
-                    </Avatar.Group>
-                  </Col>
-                </Row>
-              </Card>
-            </Col>
-          ))}
-          <Col span={24} md={12} xl={6}>
-            <Upload
-              name="avatar"
-              listType="picture-card"
-              className="avatar-uploader projects-uploader"
-              showUploadList={false}
-              action="https://www.mocky.io/v2/5cc8019d300000980a055e76"
-              beforeUpload={beforeUpload}
-              onChange={handleChange}
-            >
-              {imageURL ? (
-                <img src={imageURL} alt="avatar" style={{ width: "100%" }} />
-              ) : (
-                uploadButton
-              )}
-            </Upload>
-          </Col>
-        </Row>
-      </Card>
-    </>
-  );
+  //     <Row gutter={[24, 0]}>
+  //       <Col span={24} md={8} className="mb-24">
+  //         <Card
+  //           bordered={false}
+  //           title={<h6 className="font-semibold m-0">Thông Tin Cá Nhân</h6>}
+  //           className="header-solid h-full card-profile-information"
+  //           extra={<Button type="link">{pencil}</Button>}
+  //           bodyStyle={{ paddingTop: 0, paddingBottom: 16 }}
+  //         >
+  //           <p className="text-dark">
+  //             {" "}
+  //             Quản lý dự án với hơn 5 năm kinh nghiệm trong lĩnh vực thiết kế và chăm sóc sân vườn.{" "}
+  //           </p>
+  //           <hr className="my-25" />
+  //           <Descriptions title="Thông Tin Chi Tiết">
+  //             <Descriptions.Item label="Họ Tên" span={3}>
+  //               Nguyễn Văn A
+  //             </Descriptions.Item>
+  //             <Descriptions.Item label="Điện Thoại" span={3}>
+  //               0123456789
+  //             </Descriptions.Item>
+  //             <Descriptions.Item label="Email" span={3}>
+  //               nguyenvana@email.com
+  //             </Descriptions.Item>
+  //             <Descriptions.Item label="Địa Chỉ" span={3}>
+  //               123 Đường ABC, Quận 1, TP.HCM
+  //             </Descriptions.Item>
+  //             <Descriptions.Item label="Mạng Xã Hội" span={3}>
+  //               <a href="#pablo" className="mx-5 px-5">
+  //                 {<TwitterOutlined />}
+  //               </a>
+  //               <a href="#pablo" className="mx-5 px-5">
+  //                 {<FacebookOutlined style={{ color: "#344e86" }} />}
+  //               </a>
+  //               <a href="#pablo" className="mx-5 px-5">
+  //                 {<InstagramOutlined style={{ color: "#e1306c" }} />}
+  //               </a>
+  //             </Descriptions.Item>
+  //           </Descriptions>
+  //         </Card>
+  //       </Col>
+  //       <Col span={24} md={8} className="mb-24">
+  //         <Card
+  //           bordered={false}
+  //           title={<h6 className="font-semibold m-0">Đánh Giá Từ Khách Hàng</h6>}
+  //           className="header-solid h-full"
+  //           bodyStyle={{ paddingTop: 0, paddingBottom: 16 }}
+  //         >
+  //           <List
+  //             itemLayout="horizontal"
+  //             dataSource={data}
+  //             split={false}
+  //             className="conversations-list"
+  //             renderItem={(item) => (
+  //               <List.Item actions={[<Button type="link">XEM</Button>]}>
+  //                 <List.Item.Meta
+  //                   avatar={
+  //                     <Avatar shape="square" size={48} src={item.avatar} />
+  //                   }
+  //                   title={item.title}
+  //                   description={item.description}
+  //                 />
+  //               </List.Item>
+  //             )}
+  //           />
+  //         </Card>
+  //       </Col>
+  //       <Col span={24} md={8} className="mb-24">
+  //         <Card
+  //           bordered={false}
+  //           title={<h6 className="font-semibold m-0">Dự Án Đã Thực Hiện</h6>}
+  //           className="header-solid h-full"
+  //           bodyStyle={{ paddingTop: 0, paddingBottom: 16 }}
+  //         >
+  //           <Row gutter={[24, 24]}>
+  //             {project.map((p, index) => (
+  //               <Col span={24} md={24} key={index}>
+  //                 <Card className="card-project" bordered={false}>
+  //                   <div className="card-tag">{p.titlesub}</div>
+  //                   <h5>{p.title}</h5>
+  //                   <p>{p.disciption}</p>
+  //                   <Row gutter={[6, 0]} className="card-footer">
+  //                     <Col span={12}>
+  //                       <Button type="button">XEM CHI TIẾT</Button>
+  //                     </Col>
+  //                   </Row>
+  //                 </Card>
+  //               </Col>
+  //             ))}
+  //           </Row>
+  //         </Card>
+  //       </Col>
+  //     </Row>
+  //     <Card
+  //       bordered={false}
+  //       className="header-solid mb-24"
+  //       title={
+  //         <>
+  //           <h6 className="font-semibold">Cài Đặt Nền Tảng</h6>
+  //           <p>Quản lý cài đặt hệ thống</p>
+  //         </>
+  //       }
+  //     >
+  //       <Row gutter={[24, 24]}>
+  //         <Col span={24} md={12}>
+  //           <Card className="card-info-2" bordered={false}>
+  //             <div className="gradent">
+  //               <div className="card-content">
+  //                 <h6>Thông Báo Email</h6>
+  //                 <p>Nhận thông báo qua email</p>
+  //                 <Switch defaultChecked />
+  //               </div>
+  //             </div>
+  //           </Card>
+  //         </Col>
+  //         <Col span={24} md={12}>
+  //           <Card className="card-info-2" bordered={false}>
+  //             <div className="gradent">
+  //               <div className="card-content">
+  //                 <h6>Thông Báo SMS</h6>
+  //                 <p>Nhận thông báo qua tin nhắn</p>
+  //                 <Switch defaultChecked />
+  //               </div>
+  //             </div>
+  //           </Card>
+  //         </Col>
+  //         <Col span={24} md={12}>
+  //           <Card className="card-info-2" bordered={false}>
+  //             <div className="gradent">
+  //               <div className="card-content">
+  //                 <h6>Xác Thực 2 Lớp</h6>
+  //                 <p>Bảo mật tài khoản</p>
+  //                 <Switch />
+  //               </div>
+  //             </div>
+  //           </Card>
+  //         </Col>
+  //         <Col span={24} md={12}>
+  //           <Card className="card-info-2" bordered={false}>
+  //             <div className="gradent">
+  //               <div className="card-content">
+  //                 <h6>Sao Lưu Tự Động</h6>
+  //                 <p>Sao lưu dữ liệu định kỳ</p>
+  //                 <Switch defaultChecked />
+  //               </div>
+  //             </div>
+  //           </Card>
+  //         </Col>
+  //       </Row>
+  //     </Card>
+  //   </>
+  // );
 }
 
 export default Profile;
